% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcamCore.R
\name{run_dcam}
\alias{run_dcam}
\title{Title}
\usage{
run_dcam(
  dataset,
  penalty = "AFUSED",
  K = NA,
  K_range = 2:8,
  lambda_grid = c(0.5, 1, 2, 4, 8),
  lambda_rule = "bic",
  lambda_fixed = 2,
  params = list(gamma = 3, tau = 1, a = 3.7, eps = NA),
  reweight = "iterative",
  freeze_A = TRUE,
  max_iter = 100,
  tol = 1e-04,
  per_group = TRUE,
  max_fpr = 0.05,
  cam_init = NULL,
  verbose = TRUE,
  cam_params = list(),
  cam_seed = 2024,
  cam_dup_cor = 0.95,
  bic_zero_rule = "relative",
  bic_zero_frac = 0.001,
  bic_zero_abs = 1e-08
)
}
\description{
Title
}
