% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcamP5.R
\name{tune_lambda}
\alias{tune_lambda}
\title{Title}
\usage{
tune_lambda(
  X1,
  X2,
  S1i,
  S2i,
  A1i,
  A2i,
  penalty = "AFUSED",
  lambda_grid = c(0.5, 1, 2, 4, 8),
  params = list(),
  rule = c("bic", "truth", "fixed"),
  lambda_fixed = 2,
  truth = NULL,
  max_iter = 100,
  tol = 1e-04,
  reweight = "iterative",
  freeze_A = TRUE,
  max_fpr = 0.05,
  verbose = TRUE,
  bic_zero_rule = "relative",
  bic_zero_frac = 0.001,
  bic_zero_abs = 1e-08
)
}
\description{
Title
}
